+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Eco Crafter Toolkit Change Log
 
+## 0.8.0 (TBD)
+
+- The "Used in Recipes" tabs of item dialogs can now show recipes that aren't
+  part of the active build.
+
 ## 0.7.0 (2026-08-08)
 
 - Added support for v14 of Eco: new upgrade module system, new skills and